backend/tools.py
from typing import Optional
import os
import time
import requests
from dotenv import load_dotenv
try:
    from langchain_core.tools import tool
except ImportError:
    from langchain_core.tools import tool

load_dotenv()

AMAP_WEATHER_API = os.getenv("AMAP_WEATHER_API")
AMAP_API_KEY = os.getenv("AMAP_API_KEY")

_LAST_RAG_CONTEXT = None
_KNOWLEDGE_TOOL_CALLS_THIS_TURN = 0
_SEARCH_TOOL_CALLS_THIS_TURN = 0
_RAG_STEP_QUEUE = None
_RAG_STEP_LOOP = None   # asyncio loop, captured when setting queue
_RAG_REQUEST_ID = None


def _set_last_rag_context(context: dict):
    global _LAST_RAG_CONTEXT
    _LAST_RAG_CONTEXT = context


def get_last_rag_context(clear: bool = True) -> Optional[dict]:
    """获取最近一次 RAG 检索上下文，默认读取后清空。"""
    global _LAST_RAG_CONTEXT
    context = _LAST_RAG_CONTEXT
    if clear:
        _LAST_RAG_CONTEXT = None
    return context


def reset_tool_call_guards():
    """每轮对话开始时重置工具调用计数。"""
    global _KNOWLEDGE_TOOL_CALLS_THIS_TURN, _SEARCH_TOOL_CALLS_THIS_TURN
    _KNOWLEDGE_TOOL_CALLS_THIS_TURN = 0
    _SEARCH_TOOL_CALLS_THIS_TURN = 0


def set_rag_request_context(request_id: Optional[str]):
    global _RAG_REQUEST_ID
    _RAG_REQUEST_ID = request_id


def set_rag_step_queue(queue, request_id: Optional[str] = None):
    """设置 RAG 步骤队列，并捕获当前事件循环以便跨线程调度。"""
    global _RAG_STEP_QUEUE, _RAG_STEP_LOOP, _RAG_REQUEST_ID
    _RAG_STEP_QUEUE = queue
    _RAG_REQUEST_ID = request_id
    if queue:
        import asyncio
        try:
            _RAG_STEP_LOOP = asyncio.get_running_loop()
        except RuntimeError:
            _RAG_STEP_LOOP = asyncio.get_event_loop()
    else:
        _RAG_STEP_LOOP = None


def emit_rag_step(icon: str, label: str, detail: str = ""):
    """向队列发送一个 RAG 检索步骤。支持跨线程安全调用。"""
    global _RAG_STEP_QUEUE, _RAG_STEP_LOOP, _RAG_REQUEST_ID
    if _RAG_STEP_QUEUE is not None and _RAG_STEP_LOOP is not None:
        step = {
            "icon": icon,
            "label": label,
            "detail": detail,
            "request_id": _RAG_REQUEST_ID,
            "emitted_at_ms": round(time.perf_counter() * 1000, 2),
        }
        try:
            if not _RAG_STEP_LOOP.is_closed():
                _RAG_STEP_LOOP.call_soon_threadsafe(_RAG_STEP_QUEUE.put_nowait, step)
        except Exception:
            pass


def get_current_weather(location: str, extensions: Optional[str] = "base") -> str:
    """获取天气信息"""
    if not location:
        return "location参数不能为空"
    if extensions not in ("base", "all"):
        return "extensions参数错误，请输入base或all"

    if not AMAP_WEATHER_API or not AMAP_API_KEY:
        return "天气服务未配置（缺少 AMAP_WEATHER_API 或 AMAP_API_KEY）"

    params = {
        "key": AMAP_API_KEY,
        "city": location,
        "extensions": extensions,
        "output": "json",
    }

    try:
        resp = requests.get(AMAP_WEATHER_API, params=params, timeout=10)
        resp.raise_for_status()
        data = resp.json()
        if data.get("status") != "1":
            return f"查询失败：{data.get('info', '未知错误')}"

        if extensions == "base":
            lives = data.get("lives", [])
            if not lives:
                return f"未查询到 {location} 的天气数据"
            w = lives[0]
            return (
                f"【{w.get('city', location)} 实时天气】\n"
                f"天气状况：{w.get('weather', '未知')}\n"
                f"温度：{w.get('temperature', '未知')}℃\n"
                f"湿度：{w.get('humidity', '未知')}%\n"
                f"风向：{w.get('winddirection', '未知')}\n"
                f"风力：{w.get('windpower', '未知')}级\n"
                f"更新时间：{w.get('reporttime', '未知')}"
            )

        forecasts = data.get("forecasts", [])
        if not forecasts:
            return f"未查询到 {location} 的天气预报数据"
        f0 = forecasts[0]
        out = [f"【{f0.get('city', location)} 天气预报】", f"更新时间：{f0.get('reporttime', '未知')}", ""]
        today = (f0.get("casts") or [])[0] if f0.get("casts") else {}
        out += [
            "今日天气：",
            f"  白天：{today.get('dayweather','未知')}",
            f"  夜间：{today.get('nightweather','未知')}",
            f"  气温：{today.get('nighttemp','未知')}~{today.get('daytemp','未知')}℃",
        ]
        return "\n".join(out)

    except requests.exceptions.Timeout:
        return "错误：请求天气服务超时"
    except requests.exceptions.RequestException as e:
        return f"错误：天气服务请求失败 - {e}"
    except Exception as e:
        return f"错误：解析天气数据失败 - {e}"


@tool("search_knowledge_base")
def search_knowledge_base(query: str) -> str:
    """Search for information in the knowledge base using hybrid retrieval (dense + sparse vectors)."""
    # ... guards omitted ...
    global _KNOWLEDGE_TOOL_CALLS_THIS_TURN
    if _KNOWLEDGE_TOOL_CALLS_THIS_TURN >= 1:
        return (
            "TOOL_CALL_LIMIT_REACHED: search_knowledge_base has already been called once in this turn. "
            "Use the existing retrieval result and provide the final answer directly."
        )
    _KNOWLEDGE_TOOL_CALLS_THIS_TURN += 1

    from rag_pipeline import run_rag_graph
    rag_result = run_rag_graph(query)

    answer = rag_result.get("answer", "") if isinstance(rag_result, dict) else ""
    rag_trace = rag_result.get("rag_trace", {}) if isinstance(rag_result, dict) else {}
    if rag_trace:
        if _RAG_REQUEST_ID:
            rag_trace["request_id"] = _RAG_REQUEST_ID
        _set_last_rag_context({"rag_trace": rag_trace})

    if not answer or "don't know" in answer.lower():
        return "No relevant information found in the knowledge base to answer this question."

    return f"Knowledge Base Answer:\n{answer}"


@tool("internet_crawler_search")
def internet_crawler_search(query: str) -> str:
    """Search the internet for real-time information using an external crawler."""
    if os.getenv("DISABLE_INTERNET_CRAWLER_SEARCH", "").strip().lower() in {"1", "true", "yes", "on"}:
        return "Web search is disabled for the current evaluation run."

    global _SEARCH_TOOL_CALLS_THIS_TURN
    if _SEARCH_TOOL_CALLS_THIS_TURN >= 1:
        return "ERROR: internet_crawler_search has already been called in this turn. Use existing results or inform user if info is missing."
    _SEARCH_TOOL_CALLS_THIS_TURN += 1
    
    from duckduckgo_search import DDGS
    
    emit_rag_step("🌐", "正在联网搜索...", f"查询: {query[:50]}")
    
    try:
        with DDGS() as ddgs:
            results = list(ddgs.text(query, max_results=5))
            if not results:
                return "No search results found on the internet."
            
            formatted = []
            for i, r in enumerate(results, 1):
                title = r.get("title", "No Title")
                snippet = r.get("body", "No Body")
                link = r.get("href", "#")
                formatted.append(f"[{i}] {title}\n{snippet}\nSource: {link}")
            
            emit_rag_step("✅", "联网搜索完成", f"找到 {len(results)} 条相关结果")
            return "Web Search Results:\n\n" + "\n\n---\n\n".join(formatted)
    except Exception as e:
        error_msg = f"Web search failed: {str(e)}"
        emit_rag_step("❌", "联网搜索失败", error_msg[:100])
        return error_msg
